+++ sktime/param_est/seasonality/__init__.py
@@ -1,28 +1,11 @@
 """Module for parameter estimators of seasonality tests."""
 
-__author__ = ["fkiraly", "blazingbhavneek"]
 __all__ = [
-    "StationarityADF",
-    "StationarityKPSS",
-    "StationarityADFArch",
-    "StationarityDFGLS",
-    "StationarityPhillipsPerron",
-    "StationarityKPSSArch",
-    "StationarityZivotAndrews",
-    "StationarityVarianceRatio",
-    "BreakvarHeteroskedasticityTest",
+    "SeasonalityACFqstat",
+    "SeasonalityACF",
+    "SeasonalityPeriodogram",
 ]
 
-from sktime.param_est.stationarity._arch import (
-    StationarityADFArch,
-    StationarityDFGLS,
-    StationarityKPSSArch,
-    StationarityPhillipsPerron,
-    StationarityVarianceRatio,
-    StationarityZivotAndrews,
-)
-from sktime.param_est.stationarity._sm_breakvar import BreakvarHeteroskedasticityTest
-from sktime.param_est.stationarity._statsmodels import (
-    StationarityADF,
-    StationarityKPSS,
-)
+from sktime.param_est.seasonality._acf_qstat import SeasonalityACFqstat
+from sktime.param_est.seasonality._acf import SeasonalityACF
+from sktime.param_est.seasonality._periodogram import SeasonalityPeriodogram
